added api to upload images

+++ frontend/src/hooks/useAddToPlaylist.ts
@@ -1,14 +1,27 @@
 import axios, {AxiosError} from "axios";
 import {useCheckLike} from "@/hooks/useCheckLike.ts";
 import {useToast} from "@/components/ui/use-toast.ts";
+import {useEffect} from "react";
+import user from "@/store/user.ts";
 
-export const useAddToPlaylist = (trackId: number, playlistId: number) => {
+export const useAddToPlaylist = (trackId: number) => {
 
-    const {checkLike, isLike} = useCheckLike(playlistId, trackId);
+    useEffect(() => {
+        user.getMe();
+    }, []);
+
+    let playlistId: number;
+
+    if (user.userData.isUser) {
+        playlistId = user.me.playlists[0].id;
+    }
+
+    const {checkLike, isLike} = useCheckLike(trackId);
     const {toast} = useToast();
 
+
     const createLike = async () => {
-        if (isLike) {
+        if (isLike && user.userData.isUser) {
             await axios.delete(`http://localhost:3010/playlist/delete_track/${trackId}/${playlistId}`)
                 .then((resp) => {
                     console.log(resp.data);
@@ -36,4 +49,6 @@ export const useAddToPlaylist = (trackId: number, playlistId: number) => {
     };
 
     return {createLike};
+
+
 }
